+++ App.js
@@ -28,8 +28,7 @@ class App extends Component {
     } else if (this.props.userStore.userStatus === USER_STATUS.USER) {
       view = <Main />;
     }
-    // return <View style={styles.body}>{view}</View>;
-    return <Main />
+    return <View style={styles.body}>{view}</View>;
   }
 }
