+++ app/src/view/main/home/components/creation/NewGroupInterestsView.js
@@ -2,59 +2,55 @@ import * as React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 import { inject, observer } from 'mobx-react';
 import Colors from 'react-native/Libraries/NewAppScreen/components/Colors';
+import { TouchableOpacity } from 'react-native-gesture-handler';
 import { GROUPING_CREATION_VIEW_STATUS } from '../../../../../constant/GroupingCreationViewStatus';
 import { WINDOW_SIZE } from '../../../../../constant/WindowSize';
-import KeywordInput from './KeywordInput';
 import { COLORS } from '../../../../../assets/Colors';
 import GroupCreationProgressBar from '../GroupCreationProgressBar';
 import { FONT_SIZE } from '../../../../../constant/FontSize';
+import { EMPTY_VALUE } from '../../../../../constant/EmptyValue';
+import KeywordInput from './KeywordInput';
 
 // eslint-disable-next-line react/prop-types
-const NewGroupInterestsView = (props) => {
-  /* React.useLayoutEffect(() => {
-    // eslint-disable-next-line react/prop-types
-    props.navigation.setOptions({
-      headerRight: () => (
-        // eslint-disable-next-line react/prop-types
-        <Text onPress={() => props.navigation.navigate('InputNewGroupInterests')}>다음</Text>
-      ),
-    });
-    // eslint-disable-next-line react/prop-types,react/destructuring-assignment
-  }, [props.navigation]); */
+const NewGroupInterestsView = () => {
+  const [input, setInput] = React.useState('');
+  const [keywordList, setKeywordList] = React.useState([]);
 
-  const onHeaderNextButtonClicked = () => {
-    props.groupingCreationMainStore.groupingCreationViewChanged(
-      GROUPING_CREATION_VIEW_STATUS.DESCRIPTION
-    );
-    props.navigation.navigate('InputNewGroupMoreInfo');
+  const onkeywordInserted = (keyword) => {
+    // 공백이 빈칸이면 리턴
+    if (!keyword) {
+      return;
+    }
+    // 추가된 키워드 리스트에 추가 키워드가 있다면 리턴
+    if (keywordList.includes(keyword)) {
+      return;
+    }
+    const nextKeywords = [...keywordList, keyword];
+    setKeywordList(nextKeywords);
   };
 
-  const rightIconStyle = (groupingCreationView) => {
-    return {
-      marginRight: 15 * WINDOW_SIZE.WIDTH_WEIGHT,
-      fontSize: 18 * WINDOW_SIZE.WIDTH_WEIGHT,
-      color: props.groupingCreationMainStore.isHeaderRightIconActivated(groupingCreationView)
-        ? Colors.black
-        : '#999',
-    };
+  const onKeywordRemove = (keyword) => {
+    const nextKeywords = keywordList.filter((t) => t !== keyword);
+    setKeywordList(nextKeywords);
   };
 
-  const onKeywordChanged = (keyword) => {
-    props.groupingCreationMainStore.groupingKeywordChanged(keyword);
-
-    props.navigation.setOptions({
-      headerRight: () => (
-        <Text
-          onPress={() => {
-            onHeaderNextButtonClicked();
-          }}
-          style={rightIconStyle(GROUPING_CREATION_VIEW_STATUS.INTERESTS)}
-        >
-          다음
-        </Text>
-      ),
-    });
+  const onKeywordChange = (keywordInput) => {
+    if (keywordInput.includes(' ')) {
+      onKeywordSubmit(keywordInput);
+      return;
+    }
+    setInput(keywordInput);
   };
+
+  const onKeywordSubmit = () => {
+    console.log('submit active');
+    onkeywordInserted(input.trim());
+    setInput(EMPTY_VALUE);
+  };
+  // useEffect로 태그의 상태가 변경이 되면 다시 불러올 수 있게
+  React.useEffect(() => {
+    console.log(keywordList);
+  }, [keywordList]);
 
   return (
     // eslint-disable-next-line no-use-before-define
@@ -69,14 +65,30 @@ const NewGroupInterestsView = (props) => {
       <View style={styles.interestInputContainer}>
         <KeywordInput
           textExample="#키워드,"
-          onChangeText={onKeywordChanged.bind(this)}
-          groupingKeyword={props.groupingCreationMainStore.groupingKeyword}
+          input={input}
+          onKeywordChange={onKeywordChange}
+          keywordList={keywordList}
         />
       </View>
       <View style={styles.hotKeywordContainer}>
-        <Text style={{ fontSize: 12 * WINDOW_SIZE.HEIGHT_WEIGHT, fontWeight: 'bold' }}>
+        <Text
+          style={{
+            fontSize: 12 * WINDOW_SIZE.HEIGHT_WEIGHT,
+            fontWeight: 'bold',
+          }}
+        >
           인기 키워드
         </Text>
+      </View>
+      <View style={{ flexDirection: 'row' }}>
+        {/* keywords에서 불러오는것이 아니라 mobx에 상태에서 가져왔으면 좋겠음 */}
+        {keywordList.map((keyword) => (
+          <View style={styles.keywordListBlock}>
+            <TouchableOpacity onPress={() => onKeywordRemove(keyword)}>
+              <Text>{keyword}</Text>
+            </TouchableOpacity>
+          </View>
+        ))}
       </View>
     </View>
   );
@@ -101,6 +113,15 @@ const styles = StyleSheet.create({
   },
   hotKeywordContainer: {
     marginTop: 30 * WINDOW_SIZE.HEIGHT_WEIGHT,
+  },
+  keywordListBlock: {
+    paddingTop: 6 * WINDOW_SIZE.HEIGHT_WEIGHT,
+    paddingBottom: 6 * WINDOW_SIZE.HEIGHT_WEIGHT,
+    paddingLeft: 12 * WINDOW_SIZE.WIDTH_WEIGHT,
+    paddingRight: 12 * WINDOW_SIZE.WIDTH_WEIGHT,
+    borderRadius: 24.5 * WINDOW_SIZE.WIDTH_WEIGHT,
+    marginRight: 5 * WINDOW_SIZE.WIDTH_WEIGHT,
+    backgroundColor: COLORS.SUB_COLOR,
   },
 });
